requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
importlib_resources==6.5.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-barcode==0.16.1
python-decouple==3.8
python-escpos==3.1
PyYAML==6.0.3
qrcode==8.2
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0